Add initial Sets support for "expects" sets.
Only Sets has been modified so far, by first, returning combos of hashes with
title=>value instead of just the values. And it also appends an {"expect" => ""
} hash to each combo so something like Rules can be used to inject values in the
next commit.

This is a little rough, breaks backwards compatibility, but I think is better,
otherwise we're just trusting that the values/titles are in sync by ordering
only.