Add ExpectRule.
Sometimes a column needs a specific value that won't come from a set,
but just needs to be specifically set from a rule. This adds support for
that. The `Rule` class isn't necessary at this point, so it's gone.
There's no common behavior to be extracted from Exclude and Expect.